Moved stuff from htlc/coordinator to webutils for more re-use
